Update NexaCtrl.cpp
Only a small update to the interrupt lock to compile with more recent esp8266 "boards". Not sure at what version this wrapper was introduced. Otherwise the code compiles and work as expected on Arduino IDE 1.8.12  with ESP8266 Community 2.7.2.

Thanks for taking the time to convert this library. Saved me a bunch of time.